Fixed some warnings.
git-svn-id: file:///home/nick/svnrepo/repo/new_cyberspace/trunk@16088 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ gui_client/AddObjectDialog.cpp
@@ -84,7 +84,7 @@ AddObjectDialog::AddObjectDialog(const std::string& base_dir_path_, QSettings* s
 	models.push_back("Platonic_Solid");
 	models.push_back("Torus");
 
-	for(int i=0; i<models.size(); ++i)
+	for(size_t i=0; i<models.size(); ++i)
 	{
 		const std::string image_path = base_dir_path + "/resources/models/" + models[i] + ".png";
